requirements.txt
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
google-cloud-aiplatform==1.71.1
vertexai==1.71.1
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-wtf==1.2.1
werkzeug==2.3.7
email-validator==2.1.0
gunicorn==20.1.0
psycopg2-binary==2.9.5
Flask-Session==0.4.0
Flask-Migrate==4.0.0
